Move pkg/kubectl/drain to staging
