+++ pkg/kubectl/cmd/drain/drain_test.go
@@ -47,10 +47,10 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/cli-runtime/pkg/printers"
 	"k8s.io/client-go/rest/fake"
+	"k8s.io/kubectl/pkg/drain"
 	"k8s.io/kubectl/pkg/scheme"
 	cmdtesting "k8s.io/kubernetes/pkg/kubectl/cmd/testing"
 	cmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
-	"k8s.io/kubernetes/pkg/kubectl/drain"
 	utilpointer "k8s.io/utils/pointer"
 )
